==> pandas/01-series.ipynb <==
import pandas as pd

#Estrutura series 
#

dados = [10, 20, 30, 40, 50]
serie = pd.Series(dados)
serie

type(serie)

# Criando uma Series simples
s = pd.Series([10, 20, 30, 40], index=['a', 'b', 'c', 'd'])
print(s)
print(s['b']) 
print(s.iloc[2])

#Criando uma Series com dados de vendas mensais

vendas_mensais = pd.Series ([2000, 2500, 3000, 2800, 3500, 4000, 5000, 5110, 5800, 7000, 5500, 2600],
                            index = ['Jan', 'Fev', 'Mar', 'Abr', 'Mai', 'Jun', 'Jul', 'Ago', 'Set', 'Out', 'Nov', 'Dez'])

print(vendas_mensais)

# Calcular a média de vendas

media_vendas = vendas_mensais.mean()
media_vendas

# Mês com maior venda usando idxmax()

mes_maior_venda = vendas_mensais.idxmax()
mes_maior_venda

# Aplicar um filtro para retornar as vendas acima da média

acima_media = vendas_mensais[vendas_mensais > media_vendas] #Faz uma filtragem onde o valor da Series booleana é True
acima_media
