==> pandas/02-dataframe.ipynb <==
import pandas as pd

dados = {
    'Nome' : ['Alice', 'Bob', 'Charlie', 'Laize', 'Jessica', 'Paulo'],
    'Idade' : [24, 30, 22, 27, 31, 19],
    'Cidade' :['São Paulo', 'Rio de janeiro', 'Belo horizonte', 'São Paulo', 'Rio de janeiro', 'Belo horizonte']
}

#Transformar em um DataFrame

df = pd.DataFrame(dados)
df

#Acessar somente as 5 primeiras linhas - default
df.head()

#No paresentes dá para informar quantas linhas você deseja mostrar
df.head(2)

#Para mostrar as ultimas linhas usa-se o tail
df.tail()

#No paresentes dá para informar quantas linhas você deseja mostrar
df.tail(2)

dados_notas = {
    'Nome' : ['Alice', 'Bob', 'Charlie', 'Laize', 'Jessica', 'Paulo'],
    'Matemática' : [20, 13, 19, 18, 14, 19],
    'Física' :[17, 19, 13, 15, 17, 15],
    'Química' : [16, 17, 15, 14, 12, 16]
    }

#Dataframe das notas

df_notas = pd.DataFrame(dados_notas)
df_notas

#Calcular as médias das notas por aluno

#Criar uma nova coluna no DataFrame

df_notas['Média'] = df_notas[['Matemática', 'Física', 'Química']].mean(axis=1)
df_notas

#Identificar alunos aprovados, com nota igual ou maior que 17.

alunos_aprovados = df_notas[df_notas['Média'] >= 17]
alunos_aprovados

#Exibir apenas o nome e a média dos alunos aprovados com nota >= 17.

alunos_aprovados[['Nome', 'Média']]
