==> pandas/05-importa-csv.ipynb <==
#Site kaggle tem vários datasets disponíveis para treinar

import pandas as pd

#Importando dados de um arquivo csv

df_produtos = pd.read_csv("dados/produtos-loja.csv")
df_produtos

df_produtos.head()

df_produtos.tail()

df_produtos.describe()

#Mostra o número de linhas e o número de colunas do DataFrame

df_produtos.shape

#Verificar o tipo de dados de cada coluna

df_produtos.dtypes

#Para ordenar pela quantidade

df_produtos.sort_values(by = 'Quantidade', ascending = False)

#Contar quais categorias foram mais vendidas

df_produtos['Categoria'].value_counts()
