==> pandas/06-analise-csv.ipynb <==
import pandas as pd

df_produtos = pd.read_csv("dados/produtos-loja.csv")
df_produtos

df_produtos.dtypes

#1-Desempenho de vendas por categoria
#Criar outro dataframe com base no original sem alterar o original

vendas_categorias = df_produtos.groupby('Categoria')['Quantidade'].sum()
vendas_categorias

#2- Total de unidades vendidas

pmv = df_produtos.groupby('Produto')['Quantidade'].sum()
pmv

#3-Ordernar pelo produto mais vendido
#O pmv é uma Series e não precisa usar o argumento "by"

pmv.sort_values(ascending= False)

#4- Produto mais vendido -> .idxmax()

pmv = df_produtos.groupby('Produto')['Quantidade'].sum().idxmax()
pmv

#5- Média de preço por produto

media_preco = df_produtos.groupby('Produto')['Preço'].mean()
media_preco

#6-Correlação entre unidades vendidas e o preço unitário
#Correlacao negativa - À medida que o preço aumenta, a quantidade vendida tende a diminuir.

correlacao = df_produtos['Quantidade'].corr(df_produtos['Preço'])
correlacao

#7-Salvar o desempenho das categorias em um novo DF

vendas_categorias_df = vendas_categorias.reset_index()
vendas_categorias_df.columns = ['Categoria', 'UnidadesVendidas']
vendas_categorias_df

#8-Exportar para um arquivo CSV

vendas_categorias_df.to_csv("dados/vendas_categoria.csv", index = False)
