==> pandas/07-cria-planilha.ipynb <==
import pandas as pd

#Criar uma panilha

planetas = {
    "nome": ["Mercúrio", "Vênus", "Terra", "Marte", "Júpiter"],
    "massa_kg": [3.3e23, 4.87e24, 5.97e24, 6.42e23, 1.90e27],
    "raio_km": [2439, 6051, 6371, 3389, 69911],
    "orbita_dias": [88, 225, 365, 687, 4333],
    "tem_media_celsius": [167, 464, 15, -65, -110],
    "satelites": [0, 0, 1, 2, 79],
    "tipo": ["Rochoso", "Rochoso", "Rochoso", "Rochoso", "Gasoso"]
}


#Transformar em um DataFrame
df_planetas = pd.DataFrame(planetas)
df_planetas

#Adicionando mais uma coluna derivada das informações já existentes

df_planetas["densidade_relativa"] = df_planetas["massa_kg"] / (df_planetas["raio_km"]**3)
df_planetas

#Ordenar pelo nome
df = df_planetas.sort_values(by="nome", ascending=True)
df

#Exportar para uma planilha excel

df.to_excel("dados/planetas.xlsx", index= False)
