==> pandas/08-analise-planilha.ipynb <==
import pandas as pd

#Importar os dados da planilha excel em um DataFrame

df = pd.read_excel("dados/planetas.xlsx")
df

#Verificar os tipos de dados
df.dtypes

#Algumas medidas estatísticas

df.describe()

#Visualizar as duas primeiras linhas

df.head(2)

#Visualizar as duas últimas linhas

df.tail(2)

#Análise da planilha

planetas2 = df[['nome', 'massa_kg', 'orbita_dias', 'satelites', 'tipo']].set_index('nome')
planetas2

#Planeta com maior orbita_dias

maior_orbita = df[df['orbita_dias'] == df['orbita_dias'].max()]
maior_orbita

#Planeta com menor orbita_dias

maior_orbita = df[df['orbita_dias'] == df['orbita_dias'].min()]
maior_orbita

#Mostra a distribuição da frequência dos valores na coluna escolhida.

df['satelites'].value_counts()

#Relação entre Massa vs. Número de Satélites
#Correlação próxima a 1 significa que planetas mais massivos → mais satélites.

correlacao = df['massa_kg'].corr(df['satelites'])
correlacao
