==> pandas/11-analisa-json.ipynb <==
import json
import pandas as pd

df = pd.read_json("dados/receitas.json", encoding='utf-8')
df

#Análise total de igredientes por receita

receitas_expandidas = []

for receita in df['receitas']:
    total_igredientes = len(receita['ingredientes'])
    receita_info = {
        'titulo' : receita['titulo'],
        'tempo_preparo' : receita['tempo_preparo'],
        'porcoes' : receita['porcoes'],
        'total_igredientes' : total_igredientes
    }

    receitas_expandidas.append(receita_info)

receitas_expandidas

df_expandido = pd.DataFrame(receitas_expandidas)
df_expandido
