==> pandas/12-CSV-BD.ipynb <==
#Usar uma ORM, que é uma biblioteca que mapeia objeto-relação
#Instala a sqlalchemy


import pandas as pd
from sqlalchemy import create_engine #importou esta função do ORM

#Definir nome do banco de dados

engine = create_engine("sqlite:///loja.db")

#Criar um dataframe com o arquivo CSV

df = pd.read_csv("dados/produtos-loja.csv")
df

#Instala uma extensão SQLite viewer
#Permite visualizar o arquivo do banco de dados no VScode

try:
    df.to_sql('produtos-loja', con=engine, if_exists='replace', index=False)
    print('Dados exportados com sucesso para a tabela produtos-loja.')
except Exception as e:
    print(f'Ocorreu um erro ao exportar os daodos: {e}')

#### Análise dos dados

#1 - Total de vendas por categoria
#Usei três aspas pois o nome da minha tabela possui um hífen

consulta_1 = """SELECT Categoria, SUM(Quantidade) AS Total_Qtd_Vendida FROM "produtos-loja" GROUP BY Categoria"""
resultado_1 = pd.read_sql_query(consulta_1, engine)
resultado_1

#2 - Produto com mais vendas

consulta_2 = """SELECT Produto, SUM(Quantidade) AS Total_Qtd_Vendida FROM "produtos-loja" GROUP BY Produto ORDER BY Total_Qtd_Vendida DESC LIMIT 1"""
resultado_2 = pd.read_sql_query(consulta_2, engine)
resultado_2

#3 - Total de vendas por produto

consulta_3 = """SELECT Produto, SUM(Quantidade) AS Total_Qtd_Vendida FROM "produtos-loja" GROUP BY Produto ORDER BY Total_Qtd_Vendida DESC"""
resultado_3 = pd.read_sql_query(consulta_3, engine)
resultado_3

#4 - média de preço por produto

consulta_4 = """SELECT Produto, AVG(Preço) AS Preco_medio FROM "produtos-loja" GROUP BY Produto"""
resultado_4 = pd.read_sql_query(consulta_4, engine)
resultado_4
