==> pandas/13-transformacao_dados.ipynb <==
import pandas as pd

#Importar arquivo Json e normalizar os dados

df = pd.read_json("json/dados_hospedagem.json")
df


df.head()

#Função normalize 

df = pd.json_normalize(df["info_moveis"])
df

#A partir da terceira coluna, todos os dados estão em listas
#Precisamos retirar os itens das listas.

df.columns
list(df.columns)
columns = list(df.columns)

#Para tirar os dados das listas:

df = df.explode(columns[3:])
df


#Ajustar a indexação do índice
#inplace=True → aplica a alteração diretamente no DataFrame original, sem precisar atribuir a uma nova variável.
#drop=True → remove o índice antigo em vez de transformá-lo em uma coluna.

df.reset_index(inplace=True, drop=True)
df

#Verificar os tipos das colunas

df.dtypes

#### Conversão dados númericos

import numpy as np

#Método astype para mudar o tipo 

df["max_hospedes"] = df["max_hospedes"].astype(np.int64)
columns_numerics = ["quantidade_banheiros", "quantidade_quartos" , "quantidade_camas"]
df[columns_numerics] = df[columns_numerics].astype(np.int64)
df["avaliacao_geral"] = df["avaliacao_geral"].astype(np.float64)

float_columns = ["taxa_deposito", "taxa_limpeza", "preco"]
df[float_columns] = df[float_columns].apply(lambda col: col.str.replace("$", "").str.replace(",", ""))

df[float_columns] = df[float_columns].astype(np.float64)

df.dtypes

#### Conversão dados textuais


#Alterar todas as letras para letras minúsculas

df["descricao_local"] = df["descricao_local"].str.lower()

#Remover barras que estão no meio do texto.

df["descricao_local"] = df["descricao_local"].str.replace("[^a-zA-Z0-9\\-']", " ", regex=True)

df["descricao_local"][3000]

#Remover hífen de palavras não compostas

df["descricao_local"] = df["descricao_local"].str.replace("(?<!\\w)-(?!\\w)", " ", regex=True)

#Converter para uma lista de palavras

df["descricao_local"] = df["descricao_local"].str.split()
df

#Tirar das chaves

df["comodidades"] = df["comodidades"].str.replace("\\{|}|\\'", "", regex=True)

df["comodidades"]

#Transformar em uma lista de palavras 

df["comodidades"] = df["comodidades"].str.split(",")


df["comodidades"]
